Combine tests

check_correct checks whether a set of tests passes, and do additional, more precise tests if there were failures. The teacher should specify two code chunks;

check_code: specifies the code that checks on the (typically, final results of the) student's code. These tests are executed silently, without the reporter generating information for these.

diagnose_code: Set of tests that gets executed if the tests in check_code fail. These tests contain more detailed tests, to pinpoint the problem.

check_correct(check_code, diagnose_code)

check_or(..., incorrect_msg = NULL, choose_feedback = 1)

Arguments

check_code
In check_correct, tests that are always run.

diagnose_code
In check_correct, tests that are run only if the tests in check_code fail.

...
a set of tests passed to check_or, each set separated by a comma. Only one of these tests should pass

incorrect_msg
In check_or chracter string to override message generated by the failing test block.

choose_feedback
If all tests fail in check_or , selects the feedback of the set of tests that should be shown

Details

check_correct increases the flexibility for the student: if the tests in check_code pass, the results of the tests in diagnose_code are not considered. If you test for the end result in check_code, and only do more rigorous testing in diagnose_code, you can allow for multiple solutions to a challenge.

Similarly, check_or checks whether one of many test sets pass. That way, you can allow for multiple solutions

Examples

# NOT RUN {
# Example 1 solution code
x <- mean(1:3)

# Example SCT
check_correct({
  ex() %>% check_object("x") %>% check_equal()
}, {
  ex() %>% check_fun("mean") %>% check_arg("x") %>% check_equal()
})

# Following submissions will all be accepted:
x <- mean(1:3)
x <- 2
x <- mean(-1:5)

# Example 2 solution code
# a <- 3; b <- 4

# Example SCT
test_or(
  ex() %>% check_object('a') %>% check_equal(),
  ex() %>% check_object('b') %>% checK-equal()
)

# Following submissions will all be accepted:
a <- 3; b <- 4
a <- 3
b <- 4
# }