Run all tests for an exercise

Run all tests for an exercise and report the results (including feedback). This function is run by R Backend and should not be used by course creators.

test_exercise(sct, ex_type, pec, student_code, solution_code, student_env,
  solution_env, output_list, seed = 42)

Arguments

sct
Submission correctness tests as a character string.

ex_type
Type of the exercise

pec
pre-exercise-code

student_code
character string representing the student code

solution_code
character string representing the solution code

student_env
environment containing the objects defined by the student.

solution_env
environment containing the objects defined by solution code

output_list
the output structure that is generated by RBackend

seed
random seed that is used for SCTs that run expressions If this is the case, e.g. test_or will behave differently.

Value

A list with components passed that indicates whether all tests were sucessful, and feedback that contains a feedback message.