Test predefined R objects

At the start of your SCT, you typically want to check whether some predefined variables are still correct. test_predefined_object allows you to specify a vector of object names, together with a vector of equivalence conditions, evaluation specifications, undefined an incorrect messages. As such, test_predefined_objects is a vectorized wrapper around check_object with meaningful defeault feedback messages that tell the student to not adapt predefined objects and code in the sample code.

test_predefined_objects(name, eq_condition = "equivalent", eval = TRUE,
  undefined_msg = NULL, incorrect_msg = NULL)

Arguments

name
vector of names of the objects to check

eq_condition
vector version of eq_condition of check_equal.

eval
vector version of eval of check_equal.

undefined_msg
vector version of undefined_msg of check_object

incorrect_msg
vector version of incorrect_msg of check_object

Examples

# NOT RUN {
# Suppose the sample code specifies the variables a, b and c, 
# and you want to check that a, b and c haven't changed.
test_predefined_objects(c("a", "b", "c"))
# }