Check whether the student used the correct properties (ggvis exercises)

Check whether the student used at least as many and the correct properties as the solution inside a specific command and inside a specific function. By default, this function will compare the ggvis functions of both student and solution. However, the teacher can also state that the definition of data can be done in other functions.

test_props(index = 1, funs = "ggvis", props = NULL, allow_extra = TRUE,
  not_called_msg = NULL, incorrect_msg = NULL, error_msg = NULL)

Arguments

index
number of ggvis caommdn to be checked

funs
the function in which to look for the x and y data. If the same info is found in one function, the test passes. All the functions that the teacher specifies, must be present in the students' solution! The function only looks for properties inside the first mentioned function by the teacher.

props
set of properties to be checked. If not specified, all properties found in the solution or checked on. If specified as an empty charactor vector (c()), only the calling of the functions will be checked on.

allow_extra
whether or not the definition of additional properties is accepted (default TRUE)

not_called_msg
feedback message in case the specified function(s) was/were not found.

incorrect_msg
feedback message in case the student specified properties do not correspond with the ones in the solution.

error_msg
feedback maessage in case the student submitted a faulty ggvis call